requirements.txt
python-telegram-bot==20.3
ppyaml=6.0